refactor: Reorganized option code files
