+++ lib/src/config/config.dart
@@ -3,8 +3,11 @@ export 'config_source_provider.dart';
 export 'config_source.dart';
 export 'configuration_parser.dart';
 export 'configuration.dart';
+export 'configuration_broker.dart';
+export 'exceptions.dart';
 export 'file_system_options.dart';
 export 'multi_config_source.dart';
-export 'option_types.dart';
 export 'option_groups.dart';
+export 'option_types.dart';
+export 'options.dart';
 export 'source_type.dart';
